Normalization update
Waveguide states are no longer by default initialized as normalized states. This is to ensure compatibility with creating superpositions / beamsplitter operations. In relation to this, the twophoton initializer was adjusted to so that twophotons in the same waveguide now requires a prefactor of 1/sqrt(2) in order to be normalized.
Documentation was also updated.
Precompilation updated to be faster.